packshot-downloader/requirements.txt
aiohttp>=3.9.0,<4.0.0
Pillow>=10.3.0
tqdm>=4.66.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
brotli>=1.1.0
selenium>=4.15.0
